abstracted out config
