+++ permission-portal-cloud-functions/functions/jest.config.js
@@ -2,4 +2,5 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'jest-environment-uint8array',
   testMatch: ['**/lib/__tests__/**/*'],
+  testPathIgnorePatterns: ['config.js'],
 };
